Ficksed

+++ src/canvas.h
@@ -11,8 +11,8 @@
 	} SR_Canvas;
 
     typedef struct SR_RotatedCanvas {
-        unsigned short offset_x;
-        unsigned short offset_y;
+        int offset_x;
+        int offset_y;
         SR_Canvas canvas;
     } SR_RotateBuffer;
